Added all the questions and a button
