+++ code/src/App.js
@@ -1,59 +1,29 @@
 import React, { useState } from "react";
-import { ColorForm } from './ColorForm'
-import { GenreForm } from './GenreForm'
-
-import { Button } from './Button'
-
-const colorGroups = ["Blue", "Yellow", "Red", "Green", "Gray"];
+import { ColorForm } from './components/ColorForm'
+import { GenreForm } from './components/GenreForm'
+import { FunForm } from './components/FunForm'
+import { Button } from './components/Button'
 
 export const App = () => {
   const [colorGroup, setColorGroup] = useState();
-  const [Genre, setGenre] = useState("");
-  const [showNext, setShowNext] = useState(false);
-
-  const handleShowNext = event => {
-    event.preventDefault();
-    setShowNext(true);
-  }
+  const [genre, setGenre] = useState("");
+  const [name, setName] = useState("");
+  
   return (
 
    <div>
       <div className="firstSection">
-       {!showNext &&<form className="chooseColorForm" onSubmit={handleShowNext}>
-        <ColorForm/>
-          {colorGroups.map(group => (
-          <label key={group}>
-          {group}
-          <input
-          type="checkbox"
-          value={group}
-          onChange={event => setColorGroup(event.target.value)}
-          checked={colorGroup === group}
-        />
-       </label>
-      ))}
-      <section className="buttonContainer">
-     <Button title="Enter" />
-     </section>
-   </form>}
-    </div>
+       <form className="theForm">
+         <ColorForm name="How do you feel today?" colorGroup={colorGroup} setColorGroup={setColorGroup}/>
+         <GenreForm Genre={genre} setGenre={setGenre} />
+         <FunForm name={name} setName={setName}/>
 
-  <div className="secondSection">
-      {showNext && <GenreForm />}
-       <form className="chooseGenreForm">
-        <select
-        onChange={event => setGenre(event.target.value)}
-        value={Genre}>
-        <option value="">Choose your genre:</option>
-        <option value="Pop">Pop</option>
-        <option value="Rock">Rock</option>
-        <option value="Hiphop">Hiphop</option>
-        <option value="Country">Country</option>
-        <option value="RnB">RnB</option>
-        </select>
-    </form>
+       <section className="buttonContainer">
+         <Button title="Enter" />
+
+  </section>
+  </form>
   </div>
-  
-    </div>
+  </div>
   )
 }
