--- code/src/ColorForm.js
@@ -1,16 +1,0 @@
-import React from 'react'
- 
-export const ColorForm = () => {
-
-    return (
-        <article>
-        <h1>How do you feel today?</h1>
-        <p>Choose your color:</p>
-       </article>  
-    );
-}
-
-
-
-
-
