--- code/src/GenreForm.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-export const GenreForm = () => {
-
-    return (
-        <article>
-        <h1>What genre of music do you like the most?</h1>
-       </article> 
- 
-    );
-}
-
