+++ code/src/components/Button.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export const Button = ({title,className}) => {
+export const Button = ({title, className}) => {
     return (
     <button><p>{title}</p></button>
     )
