+++ code/src/components/FunForm.js
@@ -1,0 +1,19 @@
+import React from 'react'
+
+export const FunForm = ({name, setName}) => {
+
+    return (
+        <article>
+             <label>
+         hey
+            <input
+              type="text"
+              value={name}
+              onChange={event => setName(event.target.value)}
+            />
+          </label>
+          
+       </article> 
+ 
+    );
+}
